Centralize Zendesk URL validation rules
